將function拆開
測試從local端修改丟上去
